requirements.txt
#requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
sqlalchemy==2.0.20
pandas==2.2.3
plotly==5.15.0
requests==2.31.0
python-dotenv==1.0.0
pytest==7.4.0